InsurMan_1.2.5/requirements.txt
line-bot-sdk==1.5.0
Flask==0.12.2
gunicorn==19.7.1
beautifulsoup4==4.6.0
certifi==2017.7.27.1
chardet==3.0.4
click==6.7
idna==2.5
itsdangerous==0.24
Jinja2==2.9.6
MarkupSafe==1.0
requests==2.18.3
selenium==3.4.3
tabulate==0.7.7
urllib3==1.22
Werkzeug==0.12.2
pandas==0.22.0
